♻ Séparation des méthodes de ApiService ⬆ Mise à jour des dépendances gradle
