+++ lib/services/api/api_service.dart
@@ -1,9 +1,8 @@
 import 'dart:convert';
 
-import 'package:la_bonne_franquette_front/models/user.dart';
-import 'package:la_bonne_franquette_front/stores/secured_storage.dart';
+import 'package:la_bonne_franquette_front/services/api/api_utils_service.dart';
 
-import '../api/utils_api.dart';
+import '../../api/utils_api.dart';
 import 'package:http/http.dart' as http;
 
 class ApiService{
@@ -11,26 +10,14 @@ class ApiService{
   static final UtilsApi tool = UtilsApi();  
   static String apiQueryString = "";
 
-  static Future<String> getToken() async {
-    String? authToken = await SecuredStorage().readSecret('auth-token');
-    return authToken ?? "";
-  }
-
-  static Future<void> setBaseAddressServer() async {
-    await SecuredStorage().readSecret('adresseServeur').then((value) => {
-      apiQueryString = 'http://$value/api/v1'
-    });
-  }
-
   static Future<Map<String, String>> setHeaders(bool token) async {
     if(token){
-      String authToken = await getToken();
+      String authToken = await ApiUtilsService.getToken();
       return {
         'auth-token': authToken,
         'Content-Type': 'application/json'
       };
     }
-
     return {
       'Content-Type': 'application/json'
     };
@@ -104,46 +91,5 @@ class ApiService{
     } else {
         return true;
     }   
-  }
-
-  static Future<bool> connect({required User user}) async {
-    Map<String, String> headers = await setHeaders(false);
-    final response = await http.post(Uri.parse('$apiQueryString/auth/login'), headers: headers, body: jsonEncode(user.toJson()));
-    if(response.statusCode >= 300){
-      throw Exception('Erreur : Impossible de se connecter, ${response.statusCode} : ${response.body}');
-    } else {
-      Map<String, dynamic> token = jsonDecode(response.body);
-    SecuredStorage().writeSecrets("auth-token", token['token']);
-    return true;
-    }
-  }
-
-  static Future<String> getCacheVersion() async {
-    String token = await getToken();
-    final response = await http.get(Uri.parse('$apiQueryString/cache/version'), headers: {
-      'auth-token': token
-    });
-    if(response.statusCode >= 300){
-      throw Exception('Erreur : Impossible de récupérer la version du cache, ${response.statusCode} : ${response.body}');
-    } else {
-      return response.body.toString();
-    }    
-  }
-
-  static Future<bool> testConnection({String adresse = ""}) async {
-    http.Response response;
-    if (adresse == "") {
-      String serverAddress = await SecuredStorage().readSecret("adresseServeur") as String;
-      response = await http.get(Uri.parse("http://$serverAddress/api/v1/testConnection"));
-    } else {
-      response = await http.get(Uri.parse("http://$adresse/api/v1/testConnection"));
-    }
-
-    if (response.statusCode == 200) {
-      await ApiService.setBaseAddressServer();
-      return true;
-    } else {
-      throw Exception("Impossible de se connecter au serveur car l'adresse n'est pas configurer ou le serveur est injoignable");
-    }
   }
 }
